backend/requirements.txt
google-api-python-client==2.110.0
google-analytics-data==0.18.2
pyjwt[crypto]==2.8.0
oauth2client==4.1.3
flask-cors==4.0.0
gunicorn==21.2.0
flask==3.0.0
google-cloud==0.34.0
google-cloud-bigquery==3.16.0
flask-swagger-ui==4.11.1